Add restorePurchases to custom entitlement computation package
